requirements.txt
Flask
Flask-SQLAlchemy
Flask-Restful
alembic
marshmallow-sqlalchemy
